Oh this?
This is just a thousand different things, some maybe still in progress,
I'm not even going to mention all the stuff here but maybe more levels,
more different birds, bird bosses? shark birds? bird sharks? birds with
wheels driving around on lines, oh and the menu is completely scrapped,
now we only have something that looks like a terminal where you can run
things, including game-breaking debug-like things that I'm not sure how
to deal with yet, oh and a so called "bird monitor" that might appear
somethimes that if destroyed explodes in a fireball explosion and
destroys all birds as well as adding some stationary ones so you have
something to bounce on before more arrive, also the main.c is much
cleaner with its "state machine"-behavior, much easier to go from fade
transitions to cartoon transitions to portal transitions to a mix of
them. and also there is a cool slow motion effect for when big things
die, and maybe that is how it goes